code update

+++ src/main/java/com/elangzhi/fish/services/impl/GradeServiceImpl.java
@@ -20,7 +20,7 @@ public class GradeServiceImpl implements GradeService {
     GradeMapper gradeMapper;
 
     @Override
-    public Integer save(Grade grade) {
+    public Long save(Grade grade) {
         return null;
     }
